requirements.txt
libsass==0.21.0
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
python-frontmatter==1.0.0
PyYAML==5.1
six==1.16.0
